check exe file before compil windows

+++ build.py
@@ -358,7 +358,8 @@ def build_flutter_windows(version, features):
                   './rustdesk_portable.exe')
     print(
         f'output location: {os.path.abspath(os.curdir)}/rustdesk_portable.exe')
-    os.remove(f'./rustdesk-{version}-install.exe')
+    if os.path.exists(f'./rustdesk-{version}-install.exe'):
+        os.remove(f'./rustdesk-{version}-install.exe')
     os.rename('./rustdesk_portable.exe', f'./rustdesk-{version}-install.exe')
     print(
         f'output location: {os.path.abspath(os.curdir)}/rustdesk-{version}-install.exe')
